Fix a bug where parsers were not recognized and fix a type
Due to moving the ServiceHandler to another project, the Parsers-project was not being loaded since no direct actions on it were taken. By telling C# to load it, the parsers are found again and fix a type since our source apparently changed contents which can not fit in the defined one.